requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.6.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
bcrypt==4.1.2
python-dotenv==1.0.0
